1尝试解决多线程发送导致的冲突 备注： 发现新的问题对于文件和文件夹

+++ RemoteCtr/RemoteClient/MyTool.h
@@ -29,7 +29,7 @@ public:
 		HRESULT hRet = CreateStreamOnHGlobal(hMem, TRUE, &pStream);
 		if (hRet == S_OK) {
 			ULONG length = 0;
-			pStream->Write(pData, strBuffer.size(), &length);
+			pStream->Write(pData, (ULONG)strBuffer.size(), &length);
 			LARGE_INTEGER bg = { 0 };
 			pStream->Seek(bg, STREAM_SEEK_SET, NULL);
 			if ((HBITMAP)image != NULL)image.Destroy();
